requirements.txt
git+https://github.com/KAJdev/snowfin.git@4fddf30cfbc6fd24fbefd89843669414ffed57d2
git+https://github.com/Shinobou/hikari.git@179b822e091d5790481b06f087cc0d47a85e2081
tortoise-orm==0.19.1